Connect the DAC's AGNDF/S pins directly do the ground plane [pcb]
As investigated in #9, the DAC8831 is generating gliches in to the
AD8429 instrumentation amplifier +-7V power rails due to the ground path
for AGNDF/S pins being to long. AGNDF/S now are connected to ground
plane through a dedicated path, avoiding the top GND pour entirely.

I also rotated C81 and C82 to shorten its ground path to DGND.