Change loop compensator design for better stability [sch pcb]
Using the values calculated using the ADP1850's datasheet design
guide for C22/C32, C23/C36 and R47/R54 results in very low capacitances
and high resistance values that in turn make VS1_COMP and VS2_COMP
succeptible to noise resulting in a unstable output (see #2).

Using a scale factor of 10 for the capacitors and 1/10 for the
resistors makes VS1 and VS2 much more stable at the cost of a slower
transient responce.

I also changed L4 and L5 to 1.5uH so the peak-to-peak current is
reduced, I calculated the previous values considering the nominal
current as 10A and the max current as 15A, but we actually only operate
at 8A max, so 1.5uH is fine.