Substitute LT8362EDD with LT8362IDE [sch pcb]
LT8362 is out of stock worldwide. LT8364 is almost identical, but comes
with a 4A capable switching mosfet instead of 2A and the DFN footprint
version has 12 pins vs 10.

Due to the different footprint I reworked the -7V supply layout and did
several improvements to reduce parasitc inductances and resistances.